LPS-53627 extract ratings functionality to its own module
