+++ modules/apps/journal/journal-content-asset-addon-entries/journal-content-ratings-asset-addon-entry/bnd.bnd
@@ -1,0 +1,7 @@
+Bundle-Name: Liferay Journal Content Ratings Asset Addon Entry
+Bundle-SymbolicName: com.liferay.journal.content.asset.addon.entry.ratings
+Bundle-Version: 1.0.0
+Include-Resource:\
+	classes,\
+	META-INF/resources=src/META-INF/resources
+Web-ContextPath: /journal-content-ratings-asset-addon-entry
