+++ modules/apps/journal/journal-content-web/src/META-INF/resources/configuration.jsp
@@ -126,8 +126,6 @@ String ddmTemplateKey = journalContentDisplayContext.getDDMTemplateKey();
 			/>
 		</aui:field-wrapper>
 
-		<aui:input name="preferences--enableRatings--" type="checkbox" value="<%= journalContentDisplayContext.isEnableRatings() %>" />
-
 		<c:if test="<%= journalContentDisplayContext.isCommentsEnabled() %>">
 			<aui:input name="preferences--enableComments--" type="checkbox" value="<%= journalContentDisplayContext.isEnableComments() %>" />
